requirements.txt
flask
flask-sqlalchemy
gunicorn
pandas
plotly
python-dotenv
requests
flask-cors